+++ packages/components/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 1.67.2
+
+### Patch Changes
+
+- [#5174](https://github.com/cultureamp/kaizen-design-system/pull/5174) [`52c8fc91031f30d197359a544be14603753abaee`](https://github.com/cultureamp/kaizen-design-system/commit/52c8fc91031f30d197359a544be14603753abaee) - Fix sass deprecation warnings.
+
+  Also fixes:
+
+  - Button v1 right-to-left padding
+  - TextField reversed success icon styles
+
+## 1.67.1
+
+### Patch Changes
+
+- [#5132](https://github.com/cultureamp/kaizen-design-system/pull/5132) [`a227606160a82b80082dc5883d23a8aa9336e3d5`](https://github.com/cultureamp/kaizen-design-system/commit/a227606160a82b80082dc5883d23a8aa9336e3d5) - Update react-aria deps.
+
 ## 1.67.0
 
 ### Minor Changes